+++ internal/daemon/client.go
@@ -41,6 +41,7 @@ func NewClient() (*Client, error) {
 func (c *Client) SendNotification(
 	title,
 	body,
+	appName,
 	focusTarget,
 	focusFolder,
 	focusWindowID,
@@ -55,6 +56,7 @@ func (c *Client) SendNotification(
 		Notify: &NotifyRequest{
 			Title:              title,
 			Body:               body,
+			AppName:            appName,
 			FocusTarget:        focusTarget,
 			FocusFolder:        focusFolder,
 			FocusWindowID:      focusWindowID,